Feat : 상품 조회 #1

+++ src/main/java/com/sinse/universe/dto/response/ProductResponse.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 public record ProductResponse(
         Integer id,
         String name,
-        BigDecimal price,
+        Integer price,
         Integer stockQty,
         String categoryName,
         String artistName
